# app/schema/openapi.yaml
openapi: '3.1.1'
info:
  title: Blind Charging API
  description: |
    This API lets an application communicate with the CPL Blind Charging module via an HTTP REST API.
  version: 0.8.0
  contact:
    name: Joe Nudell
    email: jnudell@hks.harvard.edu
  license:
    name: MIT License
    url: https://opensource.org/license/mit/
...
paths:
...
  /redact:
    post:
      summary: Redact a document
      description: |
        Submit a document for redaction. Redaction happens asynchronously and may take some time.
        When finished, the redacted document will be posted to the provided callback URL.

        A callback will be POSTed to the provided URL when the redaction process is completed for each input document.
        The callback will contain either `RedactionResultSuccess` or `RedactionResultError`.
      tags:
        - redaction
      operationId: redact-documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedactionRequest"
      responses:
        "201":
          description: "Accepted"
      callbacks:
        redactionComplete:
          '{$request.body#/callbackUrl}':
            post:
              summary: Redaction complete
              description: |
                This callback is made for each input document when it is finished.
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: "#/components/schemas/RedactionResultCompleted"
              responses:
                "201":
                  description: "Accepted"
  /redact/{jurisdictionId}/{caseId}:
    get:
      summary: Get status of document redaction for a case.
      description: |
        Get the status of redaction for all documents in a case.
        This will return a list of document IDs and their redaction status.

        Generally, the push mechanism provided by the callback URL passed to the `/redact` endpoint should be used to determine when the redaction process is completed.
        However, this endpoint can be used to poll for the status of redaction if necessary.
      tags:
        - redaction
      operationId: get-redaction-status
      parameters:
        - name: jurisdictionId
          in: path
          required: true
          description: The jurisdiction ID
          schema:
            type: string
        - name: caseId
          in: path
          required: true
          description: The case ID
          schema:
            type: string
        - name: subjectId
          in: query
          required: false
          description: Optionally, filter status by a specific person.
          schema:
            type: string
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedactionStatus"
  /blindreview/{jurisdictionId}/{caseId}:
    get:
      summary: Get information about blind review for a given case.
      description: |
        This endpoint provides information about the blind review process for the given case.

        The payload will indicate whether blind review is required for this case.

        If blind review is required, this endpoint will also provide a list of redacted documents to present for review.
      tags:
        - review
      operationId: get-blind-review-info
      parameters:
        - name: jurisdictionId
          in: path
          required: true
          description: The jurisdiction ID
          schema:
            type: string
        - name: caseId
          in: path
          required: true
          description: The case ID
          schema:
            type: string
        - name: subjectId
          in: query
          required: false
          description: Optionally, the ID of a specific person
          schema:
            type: string
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlindReviewInfo"
        "424":
          description: "Documents are not processed yet"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /exposure:
    post:
      summary: Log an exposure event
      description: |
        This endpoint records which information is presented to attorneys and when, prior to them making a decision.

        Sending "exposure" events is required for all cases involved in research experiments, _both for blind review and also final review_.
      tags:
        - experiments
      operationId: log-exposure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Exposure"
      responses:
        "201":
          description: "Accepted"
  /outcome:
    post:
      summary: Log an outcome event
      description: |
        This endpoint records the charging decisions made by attorneys, both for blind review and final review.

        Sending "outcome" events is required for all cases involved in research experiments, _regardless of whether the case is subject to blind review or not_.
      tags:
        - experiments
      operationId: log-outcome
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review"
      responses:
        "201":
          description: "Accepted"
...
  /config:
    get:
      summary: Get the experiment configuration
      description: |
        Get the active randomizations configuration for the API deployment.
      tags:
        - experiments
      operationId: get-active-config
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentConfig"
    post:
      summary: Update the experiment configuration
      description: |
        Update the randomizations configuration for the API deployment.
      tags:
        - experiments
      operationId: update-config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewExperimentConfig"
      responses:
        "201":
          description: "Accepted"
        "400":
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /config/{version}:
    get:
      summary: Get the experiment configuration
      description: |
        Get the randomizations configuration for the API deployment for a specific version.
      tags:
        - experiments
      operationId: get-config
      parameters:
        - name: version
          in: path
          required: true
          description: The version of the configuration to retrieve
          schema:
            type: string
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentConfig"
  /oauth2/token:
    post:
      summary: Get an access token
      security: []
      description: |
        Get an access token to use the API.

        This endpoint is only available if the `client_credentials` flow is configured
        for the API deployment. If it is not turned on, this endpoint will return a 501.
      tags:
        - operations
      operationId: get-access-token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientCredentialsTokenRequest"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientCredentialsTokenResponse"
        "400":
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: "Not implemented"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /oauth2/revoke:
    post:
      summary: Revoke an access token
      security: []
      description: |
        Revoke an access token.

        This endpoint is only available if the `client_credentials` flow is configured
        for the API deployment. If it is not turned on, this endpoint will return a 501.
      tags:
        - operations
      operationId: revoke-access-token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientCredentialsRevokeTokenRequest"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: "Not implemented"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
security:
  - preshared: []
  - oauth2: []

# Data types
components:
...
  securitySchemes:
    preshared:
      type: http
      scheme: bearer
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth2/token
          scopes:
            default: "The only scope available at this time."
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ExperimentConfig:
      type: object
      required:
        - version
        - blob
        - active
        - createdAt
        - updatedAt
      properties:
        version:
          type: string
        blob:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        parent:
          type: string
        name:
          type: string
        description:
          type: string
        author:
          type: string
    NewExperimentConfig:
      type: object
      required:
        - blob
      properties:
        blob:
          type: string
        active:
          type: boolean
        parent:
          type: string
        name:
          type: string
        description:
          type: string
    ClientCredentialsRevokeTokenRequest:
      type: object
      required:
        - token
        - client_id
        - client_secret
      properties:
        token:
          type: string
        client_id:
          type: string
        client_secret:
          type: string
    ClientCredentialsTokenRequest:
      type: object
      required:
        - grant_type
        - client_id
        - client_secret
      properties:
        grant_type:
          type: string
          enum:
            - client_credentials
        client_id:
          type: string
        client_secret:
          type: string
    ClientCredentialsTokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
    DocumentLink:
      type: object
      required:
        - attachmentType
        - documentId
        - url
      properties:
        attachmentType:
          type: string
          enum:
            - LINK
        documentId:
          type: string
        url:
          type: string
          format: uri
    DocumentText:
      type: object
      required:
        - attachmentType
        - documentId
        - content
      properties:
        attachmentType:
          type: string
          enum:
            - TEXT
        documentId:
          type: string
        content:
          type: string
    DocumentContent:
      type: object
      required:
        - attachmentType
        - documentId
        - content
      properties:
        attachmentType:
          type: string
          enum:
            - BASE64
        documentId:
          type: string
        content:
          type: string
    DocumentJSON:
      type: object
      required:
        - attachmentType
        - documentId
        - content
      properties:
        attachmentType:
          type: string
          enum:
            - JSON
        documentId:
          type: string
        content:
          type: object
          required:
            - original
            - redacted
            - annotations
          properties:
            original:
              type: string
            redacted:
              type: string
            annotations:
              type: array
              items:
                type: object
                required:
                  - originalSpan
                  - redactedSpan
                  - valid
                  - openDelim
                  - closeDelim
                properties:
                  originalSpan:
                    type: array
                    items:
                      type: integer
                    minItems: 2
                    maxItems: 2
                  redactedSpan:
                    type: array
                    items:
                      type: integer
                    minItems: 2
                    maxItems: 2
                  valid:
                    type: boolean
                  openDelim:
                    type: [string, 'null']
                  closeDelim:
                    type: [string, 'null']
    OutputDocument:
      oneOf:
        - $ref: "#/components/schemas/DocumentLink"
        - $ref: "#/components/schemas/DocumentText"
        - $ref: "#/components/schemas/DocumentContent"
        - $ref: "#/components/schemas/DocumentJSON"
      discriminator:
        propertyName: attachmentType
        mapping:
          LINK: "#/components/schemas/DocumentLink"
          TEXT: "#/components/schemas/DocumentText"
          BASE64: "#/components/schemas/DocumentContent"
          JSON: "#/components/schemas/DocumentJSON"
    InputDocument:
      oneOf:
        - $ref: "#/components/schemas/DocumentLink"
        - $ref: "#/components/schemas/DocumentText"
        - $ref: "#/components/schemas/DocumentContent"
      discriminator:
        propertyName: attachmentType
        mapping:
          LINK: "#/components/schemas/DocumentLink"
          TEXT: "#/components/schemas/DocumentText"
          BASE64: "#/components/schemas/DocumentContent"
    BlindReviewInfo:
      type: object
      required:
        - jurisdictionId
        - caseId
        - blindReviewRequired
        - maskedSubjects
      properties:
        jurisdictionId:
          type: string
        caseId:
          type: string
        blindReviewRequired:
          type: boolean
        maskedSubjects:
          type: array
          items:
            $ref: "#/components/schemas/MaskedSubject"
    RedactionResultSuccess:
      type: object
      required:
        - jurisdictionId
        - caseId
        - maskedSubjects
        - inputDocumentId
        - redactedDocument
        - status
      properties:
        jurisdictionId:
          type: string
        caseId:
          type: string
        inputDocumentId:
          type: string
        maskedSubjects:
          type: array
          items:
            $ref: "#/components/schemas/MaskedSubject"
        redactedDocument:
          $ref: "#/components/schemas/OutputDocument"
        status:
          type: string
          enum:
            - COMPLETE
    RedactionResultError:
      type: object
      required:
        - jurisdictionId
        - caseId
        - maskedSubjects
        - inputDocumentId
        - error
        - status
      properties:
        jurisdictionId:
          type: string
        caseId:
          type: string
        inputDocumentId:
          type: string
        maskedSubjects:
          type: array
          items:
            $ref: "#/components/schemas/MaskedSubject"
        error:
          type: string
        status:
          type: string
          enum:
            - ERROR
    RedactionResultPending:
      type: object
      required:
        - jurisdictionId
        - caseId
        - maskedSubjects
        - inputDocumentId
        - status
      properties:
        jurisdictionId:
          type: string
        caseId:
          type: string
        inputDocumentId:
          type: string
        maskedSubjects:
          type: array
          items:
            $ref: "#/components/schemas/MaskedSubject"
        status:
          type: string
          enum:
            - QUEUED
            - PROCESSING
        statusDetail:
          type: string
    RedactionResultCompleted:
      description: |
        A completed redaction job. Similar to `RedactionResult`,
        but where it is not possible to see a result in an incomplete (pending or queued) state.
      oneOf:
        - $ref: "#/components/schemas/RedactionResultSuccess"
        - $ref: "#/components/schemas/RedactionResultError"
      discriminator:
        propertyName: status
        mapping:
          COMPLETE: "#/components/schemas/RedactionResultSuccess"
          ERROR: "#/components/schemas/RedactionResultError"
    RedactionResult:
      description: |
        Information about a redaction job.
      oneOf:
        - $ref: "#/components/schemas/RedactionResultSuccess"
        - $ref: "#/components/schemas/RedactionResultError"
        - $ref: "#/components/schemas/RedactionResultPending"
      discriminator:
        propertyName: status
        mapping:
          COMPLETE: "#/components/schemas/RedactionResultSuccess"
          ERROR: "#/components/schemas/RedactionResultError"
          QUEUED: "#/components/schemas/RedactionResultPending"
          PROCESSING: "#/components/schemas/RedactionResultPending"
    RedactionStatus:
      description: |
        The status of redaction for a case.
      type: object
      required:
        - jurisdictionId
        - caseId
        - requests
      properties:
        jurisdictionId:
          type: string
        caseId:
          type: string
        requests:
          type: array
          items:
            $ref: "#/components/schemas/RedactionResult"
    HumanName:
      description: A structured representation of someone's name.
      type: object
      required:
        - firstName
      properties:
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        suffix:
          type: string
        nickname:
          type: string
    MaskedSubject:
      description: Mapping between a person's ID to their alias.
      type: object
      required:
        - subjectId
        - alias
      properties:
        subjectId:
          type: string
        alias:
          type: string
    Person:
      description: Mapping between a person's ID to their name(s).
      type: object
      required:
        - subjectId
        - name
      properties:
        subjectId:
          type: string
        name:
          oneOf:
            - type: string
            - $ref: "#/components/schemas/HumanName"
        aliases:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: "#/components/schemas/HumanName"
    RedactionRequest:
      type: object
      required:
        - jurisdictionId
        - caseId
        - subjects
        - objects
      properties:
        jurisdictionId:
          type: string
        caseId:
          type: string
        subjects:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - role
              - subject
            properties:
              role:
                type: string
              subject:
                $ref: "#/components/schemas/Person"
        objects:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RedactionTarget"
        outputFormat:
          type: string
          enum:
            - PDF
            - TEXT
            - HTML
            - JSON
    RedactionTarget:
      type: object
      required:
        - document
      properties:
        document:
          $ref: "#/components/schemas/InputDocument"
        callbackUrl:
          type: string
          format: uri
        targetBlobUrl:
          type: string
          format: uri
    ReviewTimestamps:
      type: object
      required:
        - pageOpen
        - decision
      properties:
        pageOpen:
          type: string
          format: date-time
        decision:
          type: string
          format: date-time
    FinalChargeOutcome:
      type: object
      required:
        - finalChargingDecision
      properties:
        finalChargingDecision:
          description: The final charging decision.
          type: string
          enum:
            - CHARGE
            - DECLINE
          example: CHARGE
        finalChargingDecisionExplanation:
          type: string
          maxLength: 4194303
          example: "The accused was caught on camera with the stolen goods."
    BlindDecisionOutcome:
      type: object
      required:
        - outcomeType
        - blindChargingDecision
      properties:
        outcomeType:
          type: string
          enum:
            - BLIND_DECISION
          example: BLIND_DECISION
        blindChargingDecision:
          description: The likely charging decision after blind review.
          type: string
          enum:
            - CHARGE_LIKELY
            - CHARGE_MAYBE
            - DECLINE_MAYBE
            - DECLINE_LIKELY
          example: CHARGE_LIKELY
        blindChargingDecisionExplanation:
          type: string
          maxLength: 4194303
          example: "The accused was caught on camera with the stolen goods."
        additionalEvidence:
          type: string
          maxLength: 4194303
          example: "The accused has a history of theft."
    DisqualifyingReason:
      type: string
      enum:
        - ASSIGNED_TO_UNBLIND
        - CASE_TYPE_INELIGIBLE
        - PRIOR_KNOWLEDGE_BIAS
        - NARRATIVE_INCOMPLETE
        - REDACTION_MISSING
        - REDACTION_ILLEGIBLE
        - OTHER
    DisqualifyOutcome:
      type: object
      required:
        - outcomeType
        - disqualifyingReason
      properties:
        outcomeType:
          type: string
          enum:
            - DISQUALIFICATION
          example: DISQUALIFICATION
        disqualifyingReason:
          oneOf:
            - $ref: "#/components/schemas/DisqualifyingReason"
              example: PRIOR_KNOWLEDGE_BIAS
            - type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/DisqualifyingReason"
              example: [PRIOR_KNOWLEDGE_BIAS, OTHER]
        disqualifyingReasonExplanation:
          type: string
          maxLength: 4194303
          example: "I have prior knowledge of the individuals involved in this case."
    ReviewProtocol:
      type: string
      enum:
        - BLIND_REVIEW
        - FINAL_REVIEW
    BlindReviewDecision:
      type: object
      required:
        - protocol
        - outcome
      properties:
        protocol:
          type: string
          enum:
            - BLIND_REVIEW
          example: BLIND_REVIEW
        outcome:
          oneOf:
            - $ref: "#/components/schemas/BlindDecisionOutcome"
            - $ref: "#/components/schemas/DisqualifyOutcome"
          discriminator:
            propertyName: outcomeType
            mapping:
              BLIND_DECISION: "#/components/schemas/BlindDecisionOutcome"
              DISQUALIFICATION: "#/components/schemas/DisqualifyOutcome"
    FinalReviewDecision:
      type: object
      required:
        - protocol
        - outcome
      properties:
        protocol:
          type: string
          enum:
            - FINAL_REVIEW
          example: FINAL_REVIEW
        outcome:
          $ref: "#/components/schemas/FinalChargeOutcome"
    ReviewDecision:
      oneOf:
        - $ref: "#/components/schemas/BlindReviewDecision"
        - $ref: "#/components/schemas/FinalReviewDecision"
      discriminator:
        propertyName: protocol
        mapping:
          BLIND_REVIEW: "#/components/schemas/BlindReviewDecision"
          FINAL_REVIEW: "#/components/schemas/FinalReviewDecision"
    Exposure:
      type: object
      required:
        - jurisdictionId
        - caseId
        - subjectId
        - reviewingAttorneyMaskedId
        - documentIds
        - protocol
      properties:
        jurisdictionId:
          type: string
        caseId:
          type: string
        subjectId:
          oneOf:
            - type: string
            - type: array
              minItems: 1
              items:
                type: string
        reviewingAttorneyMaskedId:
          type: string
        documentIds:
          type: array
          minItems: 1
          items:
            type: string
        protocol:
          $ref: "#/components/schemas/ReviewProtocol"
        extra:
          type: string
          description: |
            Additional information about the exposure event.
            This field is optional and can be used to provide additional context.
    Review:
      type: object
      required:
        - jurisdictionId
        - caseId
        - subjectId
        - reviewingAttorneyMaskedId
        - documentIds
        - protocol
        - decision
        - timestamps
      properties:
        jurisdictionId:
          type: string
        caseId:
          type: string
        subjectId:
          oneOf:
            - type: string
            - type: array
              minItems: 1
              items:
                type: string
        reviewingAttorneyMaskedId:
          type: string
        documentIds:
          type: array
          minItems: 1
          items:
            type: string
        decision:
          $ref: "#/components/schemas/ReviewDecision"
        timestamps:
          $ref: "#/components/schemas/ReviewTimestamps"
...